backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.45.0
django-cors-headers==3.10.0
django-filter==21.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2==2.9.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
python3-openid==3.2.0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wincertstore==0.2
